Factor out base-conversation-handlers.tsx, parsing.ts.
